requirements.txt
Django
requests
google-generativeai
python-dotenv
